refactor: 렌더링 함수 분리
